+++ CMakeLists.txt
@@ -1165,7 +1165,7 @@ add_library(llama
             llama.h
             unicode.h
             unicode.cpp
-            unicodedata.cpp
+            unicode-data.cpp
             )
 
 target_include_directories(llama PUBLIC .)
